clean up code base and add more basic functions
Now there is support for basic checking of tasks files. Also, this adds
logging via console.log to give our users a hint, what's going on.

The test.js is working and if all went well, you should be able to see
"It works!" in your browser, if requesting http://localhost:3000